+++ sdk/schemaregistry/azure-schemaregistry-avroserializer/azure/schemaregistry/serializer/avroserializer/aio/_schema_registry_avro_serializer_async.py
@@ -23,10 +23,6 @@
 # IN THE SOFTWARE.
 #
 # --------------------------------------------------------------------------
-try:
-    from functools import lru_cache
-except ImportError:
-    from backports.functools_lru_cache import lru_cache
 from io import BytesIO
 from typing import Any, Dict, Mapping
 import avro
@@ -34,6 +30,7 @@ import avro
 from ._async_lru import alru_cache
 from .._constants import SCHEMA_ID_START_INDEX, SCHEMA_ID_LENGTH, DATA_START_INDEX
 from .._avro_serializer import AvroObjectSerializer
+from .._utils import parse_schema
 
 
 class AvroSerializer(object):
@@ -43,10 +40,10 @@ class AvroSerializer(object):
 
     :keyword client: Required. The schema registry client
      which is used to register schema and retrieve schema from the service.
-    :paramtype client: ~azure.schemaregistry.SchemaRegistryClient
+    :paramtype client: ~azure.schemaregistry.aio.SchemaRegistryClient
     :keyword str group_name: Required. Schema group under which schema should be registered.
     :keyword bool auto_register_schemas: When true, register new schemas passed to serialize.
-     Otherwise, and by default, fail if it has not been pre-registered in the registry.
+     Otherwise, and by default, serialization will fail if the schema has not been pre-registered in the registry.
 
     """
 
@@ -72,7 +69,7 @@ class AvroSerializer(object):
 
     async def __aexit__(self, *exc_details):
         # type: (Any) -> None
-        await self._schema_registry_client.__exit__(*exc_details)
+        await self._schema_registry_client.__aexit__(*exc_details)
 
     async def close(self):
         # type: () -> None
@@ -81,7 +78,7 @@ class AvroSerializer(object):
         """
         await self._schema_registry_client.close()
 
-    @alru_cache(maxsize=128)
+    @alru_cache(maxsize=128, cache_exceptions=False)
     async def _get_schema_id(self, schema_name, schema_str, **kwargs):
         # type: (str, str, Any) -> str
         """
@@ -90,8 +87,7 @@ class AvroSerializer(object):
 
         :param schema_name: Name of the schema
         :type schema_name: str
-        :param schema: Schema object
-        :type schema: avro.schema.Schema
+        :param str schema_str: Schema string
         :return: Schema Id
         :rtype: str
         """
@@ -100,7 +96,7 @@ class AvroSerializer(object):
         )
         return schema_properties.id
 
-    @alru_cache(maxsize=128)
+    @alru_cache(maxsize=128, cache_exceptions=False)
     async def _get_schema(self, schema_id, **kwargs):
         # type: (str, Any) -> str
         """
@@ -114,11 +110,6 @@ class AvroSerializer(object):
             schema_id, **kwargs
         )
         return schema.schema_definition
-
-    @classmethod
-    @lru_cache(maxsize=128)
-    def _parse_schema(cls, schema):
-        return avro.schema.parse(schema)
 
     async def serialize(self, value, **kwargs):
         # type: (Mapping[str, Any], Any) -> bytes
@@ -138,7 +129,7 @@ class AvroSerializer(object):
         except KeyError as e:
             raise TypeError("'{}' is a required keyword.".format(e.args[0]))
 
-        cached_schema = AvroSerializer._parse_schema(raw_input_schema)
+        cached_schema = parse_schema(raw_input_schema)
         record_format_identifier = b"\0\0\0\0"
         schema_id = await self._get_schema_id(cached_schema.fullname, str(cached_schema), **kwargs)
         data_bytes = self._avro_serializer.serialize(value, cached_schema)