+++ sdk/schemaregistry/azure-schemaregistry-avroserializer/samples/async_samples/eventhub_receive_integration.py
@@ -31,10 +31,11 @@ eventhub_consumer = EventHubConsumerClient.from_connection_string(
     eventhub_name=EVENTHUB_NAME,
 )
 # create a AvroSerializer instance
+azure_credential = DefaultAzureCredential()
 avro_serializer = AvroSerializer(
     client=SchemaRegistryClient(
         fully_qualified_namespace=SCHEMAREGISTRY_FULLY_QUALIFIED_NAMESPACE,
-        credential=DefaultAzureCredential()
+        credential=azure_credential
     ),
     group_name=GROUP_NAME,
     auto_register_schemas=True
@@ -62,6 +63,10 @@ async def main():
             )
     except KeyboardInterrupt:
         print('Stopped receiving.')
+    finally:
+        await avro_serializer.close()
+        await azure_credential.close()
+        await eventhub_consumer.close()
 
 if __name__ == '__main__':
     loop = asyncio.get_event_loop()