+++ sdk/schemaregistry/azure-schemaregistry-avroserializer/tests/conftest.py
@@ -1,0 +1,16 @@
+# ------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# -------------------------------------------------------------------------
+
+import os
+import sys
+import uuid
+
+import pytest
+
+# Ignore async tests for Python < 3.5
+collect_ignore = []
+if sys.version_info < (3, 5):
+    collect_ignore.append("*_async.py")